+++ src/matrix_math.rs
@@ -1,17 +1,17 @@
 extern crate nalgebra as na;
 use na::{DMat};
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 use std::cmp::Ordering;
 use std::ops::*;
 
 
-pub fn div_matrix_by_matrix_each_value(mat1: &DMat<f32>, mat2: &DMat<f32>) -> DMat<f32> {
+pub fn div_matrix_by_matrix_each_value(mat1: &DMat<f64>, mat2: &DMat<f64>) -> DMat<f64> {
    assert_eq!(mat1.nrows(), mat2.nrows());
    assert_eq!(mat1.ncols(), mat2.ncols());
    
    let rows_size = mat1.nrows();
    let cols_size = mat1.ncols();
-   let mut result: DMat<f32> = DMat::new_zeros(rows_size, cols_size);
+   let mut result: DMat<f64> = DMat::new_zeros(rows_size, cols_size);
 
    for r in 0..rows_size {
 	for c in 0..cols_size {
@@ -22,13 +22,13 @@ pub fn div_matrix_by_matrix_each_value(mat1: &DMat<f32>, mat2: &DMat<f32>) -> DM
   return result;
 }
 
-pub fn mul_matrix_by_matrix_each_value(mat1: &DMat<f32>, mat2: &DMat<f32>) -> DMat<f32> {
+pub fn mul_matrix_by_matrix_each_value(mat1: &DMat<f64>, mat2: &DMat<f64>) -> DMat<f64> {
    assert_eq!(mat1.nrows(), mat2.nrows());
    assert_eq!(mat1.ncols(), mat2.ncols());
    
    let rows_size = mat1.nrows();
    let cols_size = mat1.ncols();
-   let mut result: DMat<f32> = DMat::new_zeros(rows_size, cols_size);
+   let mut result: DMat<f64> = DMat::new_zeros(rows_size, cols_size);
 
    for r in 0..rows_size {
 	for c in 0..cols_size {
@@ -39,7 +39,7 @@ pub fn mul_matrix_by_matrix_each_value(mat1: &DMat<f32>, mat2: &DMat<f32>) -> DM
   return result;
 }
 
-pub fn prod(vec : &[f32]) -> f32 {
+pub fn prod(vec : &[f64]) -> f64 {
 	let mut result = 1.0;
 	
 	for x in vec {
@@ -49,10 +49,10 @@ pub fn prod(vec : &[f32]) -> f32 {
 	return result;
 }
 
-pub fn div_scalar_by_matrix(scalar: f32, mat: &DMat<f32>) -> DMat<f32>{
+pub fn div_scalar_by_matrix(scalar: f64, mat: &DMat<f64>) -> DMat<f64>{
    let rows_size = mat.nrows();
    let cols_size = mat.ncols();
-   let mut result: DMat<f32> = DMat::new_zeros(rows_size, cols_size);
+   let mut result: DMat<f64> = DMat::new_zeros(rows_size, cols_size);
 
    for r in 0..rows_size {
 	for c in 0..cols_size {
@@ -64,10 +64,10 @@ pub fn div_scalar_by_matrix(scalar: f32, mat: &DMat<f32>) -> DMat<f32>{
 }
 
 
-pub fn power_matrix_by_scalar(mat: &DMat<f32>, s: i32) -> DMat<f32> {
+pub fn power_matrix_by_scalar(mat: &DMat<f64>, s: i32) -> DMat<f64> {
    let rows_size = mat.nrows();
    let cols_size = mat.ncols();
-   let mut result: DMat<f32> = DMat::new_zeros(rows_size, cols_size);
+   let mut result: DMat<f64> = DMat::new_zeros(rows_size, cols_size);
  
    for r in 0..rows_size {
 	for c in 0..cols_size {
@@ -78,8 +78,8 @@ pub fn power_matrix_by_scalar(mat: &DMat<f32>, s: i32) -> DMat<f32> {
   return result;
 }
 
-pub fn pow(x: f32, s: i32) -> f32{
-  let mut result: f32 = 1.0;
+pub fn pow(x: f64, s: i32) -> f64{
+  let mut result: f64 = 1.0;
 
   for _ in 0..s {
 	result = result * x;
@@ -87,11 +87,11 @@ pub fn pow(x: f32, s: i32) -> f32{
   return result;
 }
 
-pub fn matrix_sqrt(mat : &DMat<f32>) -> DMat<f32> {
+pub fn matrix_sqrt(mat : &DMat<f64>) -> DMat<f64> {
 	let nrows = mat.nrows();
 	let ncols = mat.ncols();
 	
-	let mut result : DMat<f32> = DMat::new_zeros(nrows, ncols);
+	let mut result : DMat<f64> = DMat::new_zeros(nrows, ncols);
 	
 	for r in 0..nrows {
 		for c in 0..ncols {
@@ -102,7 +102,7 @@ pub fn matrix_sqrt(mat : &DMat<f32>) -> DMat<f32> {
 	return result;
 }
 
-pub fn mat_map<F>(mat_src: &na::DMat<f32>, mut f: F) -> na::DMat<f32> where F: FnMut(f32) -> f32 {
+pub fn mat_map<F>(mat_src: &na::DMat<f64>, mut f: F) -> na::DMat<f64> where F: FnMut(f64) -> f64 {
     let mut mat = na::DMat::new_zeros(mat_src.nrows(), mat_src.ncols());
     for i in 0..mat.nrows() {
         for j in 0..mat.ncols() {
@@ -112,7 +112,7 @@ pub fn mat_map<F>(mat_src: &na::DMat<f32>, mut f: F) -> na::DMat<f32> where F: F
     mat
 }
 
-pub fn mat_map_mut<F>(mut mat: na::DMat<f32>, mut f: F) -> na::DMat<f32> where F: FnMut(f32) -> f32 {
+pub fn mat_map_mut<F>(mut mat: na::DMat<f64>, mut f: F) -> na::DMat<f64> where F: FnMut(f64) -> f64 {
     for i in 0..mat.nrows() {
         for j in 0..mat.ncols() {
             mat[(i, j)] = f(mat[(i, j)]);
